refactor: extracts import repository to its own module
