+++ api/apps/api/src/modules/clone/import/application/import-repository/memory--import.repository.adapter.ts
@@ -1,0 +1,28 @@
+import {
+  Failure as RepoFailure,
+  ImportRepository,
+  Success,
+} from '@marxan-api/modules/clone/import/application/import-repository/import.repository.port';
+import {
+  Import,
+  ImportId,
+  ImportSnapshot,
+} from '@marxan-api/modules/clone/import';
+import { Either, right } from 'fp-ts/Either';
+
+export class MemoryImportRepository extends ImportRepository {
+  entities: Record<string, ImportSnapshot> = {};
+
+  async find(importId: ImportId): Promise<Import | undefined> {
+    const snapshot = this.entities[importId.value];
+    if (!snapshot) return;
+
+    return Import.from(snapshot);
+  }
+
+  async save(importRequest: Import): Promise<Either<RepoFailure, Success>> {
+    const snapshot = importRequest.toSnapshot();
+    this.entities[snapshot.id.value] = snapshot;
+    return right(true);
+  }
+}
